- Property Formatting sample updated
