+++ src/frontend-samples/image-export/ImageExportWidget.tsx
@@ -11,7 +11,6 @@ import "./ImageExport.scss";
 
 export const ImageExportWidget: React.FunctionComponent = () => {
 
-  // Display drawing and sheet options in separate sections.
   return (
     <>
       <div className="sample-options">
