+++ src/frontend-samples/property-formatting-sample/PropertyFormatting.scss
@@ -2,6 +2,18 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
+.sample-options {
+  color: var(--buic-foreground-body);
+  padding: 8px;
+}
+
+.sample-options-2col {
+  display: grid;
+  row-gap: 8px;
+  column-gap: 8px;
+  grid-template-columns: 1fr 2fr;
+  align-items: center;
+}
 .table-box {
   display: flex;
   height: 300px;
