+++ src/frontend-samples/reality-data-sample/RealityDataWidget.tsx
@@ -68,7 +68,7 @@ const RealityDataWidget: React.FunctionComponent = () => {
     );
   };
 
-  // Display drawing and sheet options in separate sections.
+
   return (
     <>
       <div className="sample-options">
