+++ src/frontend-samples/zoom-to-elements-sample/ZoomToElementsWidget.tsx
@@ -92,7 +92,6 @@ const ZoomToElementsWidget: React.FunctionComponent = () => {
     );
   };
 
-  // Display drawing and sheet options in separate sections.
   return (
     <>
       <div className="sample-options">
